refactor(iso): configure ingest tooling
Extract the reusable ingest logic away from the CLI entrypoint so tests can import it without writing generated data. Add environment-based ISO fetch and ingest configuration for local update workflows.
