+++ .gitignore
@@ -1,4 +1,5 @@
 dist/
+.env
 
 node_modules
 package-lock.json
