+++ env.example
@@ -1,0 +1,6 @@
+# ISO 4217 Currency Data Configuration
+# Copy this file to .env and adjust values as needed
+
+ISO_4217_URL=https://www.six-group.com/dam/download/financial-information/data-center/iso-currrency/lists/list-one.xml
+ISO_4217_XML_PATH=resources/iso-4217/list-one.xml
+ISO_4217_JSON_PATH=data/iso-4217/list-one.json
