+++ package.json
@@ -32,6 +32,7 @@
     "@types/node": "^20.0.0",
     "@vitest/ui": "^3.2.4",
     "axios": "^0.19.2",
+    "dotenv": "^17.2.2",
     "fast-xml-parser": "^4.3.0",
     "tsup": "^8.0.0",
     "tsx": "^4.20.5",
