+++ scripts/iso-4217/fetch.ts
@@ -1,5 +1,6 @@
 import * as fs from 'node:fs';
 import axios from 'axios';
+import { ISO_4217_CONFIG } from '../../config';
 
 async function download(url: string, filePath: string): Promise<void> {
   const writer = fs.createWriteStream(filePath);
@@ -19,12 +20,11 @@ async function download(url: string, filePath: string): Promise<void> {
 }
 
 async function downloadIso(): Promise<void> {
-  const url = 'https://www.six-group.com/dam/download/financial-information/data-center/iso-currrency/lists/list-one.xml';
-  const filePath = 'resources/iso-4217/list-one.xml';
+  const { url, xmlPath } = ISO_4217_CONFIG;
 
   try {
-    await download(url, filePath);
-    console.log(`Downloaded ${url} to ${filePath}`);
+    await download(url, xmlPath);
+    console.log(`Downloaded ${url} to ${xmlPath}`);
   } catch (e) {
     console.error(`Error downloading ${url}`);
     console.error(e);
